+++ taf/api/dependencies.py
@@ -3,6 +3,7 @@ from logging import DEBUG, ERROR
 from typing import Dict, Optional
 import click
 from taf.api.targets import register_target_files
+from taf.api.utils._conf import read_keys_name_mapping
 import taf.repositoriesdb as repositoriesdb
 from logdecorator import log_on_end, log_on_error, log_on_start
 from taf.api.utils._git import check_if_clean_and_synced
@@ -74,6 +75,7 @@ def add_dependency(
     commit: Optional[bool] = True,
     push: Optional[bool] = True,
     no_prompt: Optional[bool] = False,
+    keys_description: Optional[str] = None,
 ) -> None:
     """
     Add a dependency (an authentication repository) to dependencies.json or update it if it was already added to this file.
@@ -110,6 +112,8 @@ def add_dependency(
         raise TAFError("Authentication repository's path not provided")
 
     auth_repo = AuthenticationRepository(path=path, pin_manager=pin_manager)
+    keys_name_mappings = read_keys_name_mapping(keys_description)
+    auth_repo.add_key_names(keys_name_mappings)
     if not auth_repo.is_git_repository_root:
         taf_logger.error(f"{path} is not a git repository!")
         return
@@ -200,6 +204,7 @@ def remove_dependency(
     prompt_for_keys: Optional[bool] = False,
     commit: Optional[bool] = True,
     push: Optional[bool] = True,
+    keys_description: Optional[str] = None,
 ) -> None:
     """
     Remove a dependency (an authentication repository) from dependencies.json
@@ -223,6 +228,8 @@ def remove_dependency(
         raise TAFError("Authentication repository's path not provided")
 
     auth_repo = AuthenticationRepository(path=path, pin_manager=pin_manager)
+    keys_name_mappings = read_keys_name_mapping(keys_description)
+    auth_repo.add_key_names(keys_name_mappings)
     if not auth_repo.is_git_repository_root:
         print(f"{path} is not a git repository!")
         return